Added RSpecs - https://dchealthlink.kanbanize.com/ctrl_board/15/cards/4193/details
